// src/interface.ts
export interface IComparingTemp {
  id: string;
  name: string;
  type: string;
  message: string;
  pages: number;
  pagedata: string[][];
  score?: number;
}
// ...
export interface IApiSigboxTemplate {
  //for templates
  left: string;
  top: string;
  width: string;
  height: string;
  esigntype: string;
  boxid: string;
  pagenumber: string;
  font: string;
  fontsize: string;
  lineheight: string;
  fontcolor: string;
  signsetid: string;
  fieldname: string;
  fieldvalue: string;
  fieldlabel: string;
  defaultvalue: string;
  depfield: string;
  depfieldvalue: string;
  depoperator: string;
  fieldrequired: string;
  checkedvalue: string;
  uncheckedvalue: string;
  role: string;
  signerid: number;
  page: number;
  xp: number;
  yp: number;
  hp: number;
  wp: number;
  type: number;
  text: string;
  checked: boolean;
  required: boolean;
}
// ...
export interface IPendingDoc {
  docid: string;
  docname: string;
  createdon: string;
}
// ...
export interface IApiUser {
  controlid: string;
  username: string;
  userid: string;
  email: string;
  phonenumber: string;
  fullname: string;
  session: string;
  sessionlife: number;
  retentionperiod: number;
  availablemfoptions: string;
  multifactor: string;
  controlids: string[];
  permissions: string[];
}